lost-and-found-app/user_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pyjwt
python-dotenv